# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.createServiceOrderRequest:
    properties:
      client_document_number:
        type: string
      client_document_type:
        type: string
      client_email:
        type: string
      client_name:
        type: string
      client_phone:
        type: string
      customer_complaint:
        type: string
      parts:
        items:
          $ref: '#/definitions/serviceorder.ItemInput'
        type: array
      services:
        items:
          $ref: '#/definitions/serviceorder.ItemInput'
        type: array
      vehicle_brand:
        type: string
      vehicle_color:
        type: string
      vehicle_manufacture_year:
        type: integer
      vehicle_model:
        type: string
      vehicle_model_year:
        type: integer
      vehicle_plate:
        type: string
    required:
    - client_document_number
    - client_document_type
    - client_name
    - vehicle_brand
    - vehicle_model
    - vehicle_model_year
    - vehicle_plate
    type: object
  controllers.createServiceOrderResponse:
    properties:
      budget_id:
        type: string
      budget_status:
        type: string
      code:
        type: string
      service_order_id:
        type: string
      total_cents:
        type: integer
    type: object
...
  controllers.serviceOrderDetailResponse:
    properties:
      client_id:
        type: string
      code:
        type: string
      id:
        type: string
      opened_at:
        type: string
      status:
        type: string
      vehicle_id:
        type: string
    type: object
  controllers.serviceOrderSummaryResponse:
    properties:
      client_id:
        type: string
      client_name:
        type: string
      code:
        type: string
      id:
        type: string
      opened_at:
        type: string
      plate:
        type: string
      status:
        type: string
      vehicle_id:
        type: string
    type: object
  dto.AdjustStockRequest:
    properties:
      movement_type:
        type: string
      notes:
        type: string
      quantity:
        type: integer
    required:
    - movement_type
    - quantity
    type: object
  dto.ClientResponse:
    properties:
      document_number:
        type: string
      document_type:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  dto.CreateClientRequest:
    properties:
      document_number:
        type: string
      document_type:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    required:
    - document_number
    - document_type
    - name
    type: object
  dto.CreatePartRequest:
    properties:
      active:
        type: boolean
      description:
        type: string
      name:
        type: string
      sku:
        type: string
      stock_quantity:
        type: integer
      unit_price_cents:
        type: integer
    required:
    - name
    - sku
    type: object
...
  dto.PartResponse:
    properties:
      active:
        type: boolean
      description:
        type: string
      id:
        type: string
      name:
        type: string
      sku:
        type: string
      stock_quantity:
        type: integer
      unit_price_cents:
        type: integer
    type: object
...
  serviceorder.ItemInput:
    properties:
      id:
        type: string
      quantity:
        type: integer
    type: object
info:
  contact: {}
  description: API for the Tech Challenge project.
  title: Tech Challenge API
  version: "1.0"
paths:
  /admin/clients:
    get:
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ClientResponse'
            type: array
      summary: List clients
      tags:
      - admin-clients
    post:
      parameters:
      - description: Client
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateClientRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ClientResponse'
      summary: Create client
      tags:
      - admin-clients
  /admin/clients/{id}:
    delete:
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete client
      tags:
      - admin-clients
    get:
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ClientResponse'
      summary: Get client
      tags:
      - admin-clients
    put:
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      - description: Client
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateClientRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ClientResponse'
      summary: Update client
      tags:
      - admin-clients
...
  /admin/parts:
    get:
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PartResponse'
            type: array
      summary: List parts and supplies
      tags:
      - admin-parts
    post:
      parameters:
      - description: Part
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePartRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PartResponse'
      summary: Create part
      tags:
      - admin-parts
  /admin/parts/{id}:
    delete:
      parameters:
      - description: Part ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete part
      tags:
      - admin-parts
    get:
      parameters:
      - description: Part ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PartResponse'
      summary: Get part
      tags:
      - admin-parts
    put:
      parameters:
      - description: Part ID
        in: path
        name: id
        required: true
        type: string
      - description: Part
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePartRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PartResponse'
      summary: Update part
      tags:
      - admin-parts
  /admin/parts/{id}/stock-movements:
    post:
      parameters:
      - description: Part ID
        in: path
        name: id
        required: true
        type: string
      - description: Stock movement
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdjustStockRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PartResponse'
      summary: Adjust part stock (creates stock movement)
      tags:
      - admin-parts
  /admin/service-orders:
    get:
      parameters:
      - description: Status
        in: query
        name: status
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.serviceOrderSummaryResponse'
            type: array
      security:
      - BearerAuth: []
      summary: List service orders
      tags:
      - admin-service-orders
    post:
      parameters:
      - description: Service order
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.createServiceOrderRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.createServiceOrderResponse'
      summary: Create service order (draft budget)
      tags:
      - admin-service-orders
  /admin/service-orders/{id}:
    get:
      parameters:
      - description: Service Order ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.serviceOrderDetailResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get service order
      tags:
      - admin-service-orders
  /admin/service-orders/{id}/budget/send:
    post:
      parameters:
      - description: Service Order ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Send budget for approval
      tags:
      - admin-service-orders
...
  /admin/service-orders/{id}/diagnosis/start:
    post:
      parameters:
      - description: Service Order ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Start diagnosis
      tags:
      - admin-service-orders
  /admin/service-orders/{id}/finish:
    post:
      parameters:
      - description: Service Order ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Finish service order
      tags:
      - admin-service-orders
...
securityDefinitions:
  BearerAuth:
    description: Type `Bearer ` followed by your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"